Move LLM handler to a module
